Move search page to /pages section.
It was implemented as template. However, there is no need. We can
safely write it as a page and it will get picked up by iteration of
PAGES automatically. Further, this makes debugging much easier that we
don't need to reload template all the time. Third benefit is that it
will always stay available in page header as designed.